Add the "init" node mon state
When a daemon stops it announces its "maintenance" or "upgrade" state, that
peers interpret as "don't drop this node's data until it's ready to send
fresh data (to not trigger failback during maintenance or upgrade).

As soon as the daemon starts, it announces it's no longer in "maintenance" or
"upgrade", but the service status are still not re-evaluated, so peer nodes
can transition services to "ready" in preparation of a failover.

If the status re-eval lasts more than the ready-period, the failover will
start, which is not the intended behaviour.

This patch set the node mon state to "init" on startup, which peers interpret
as "don't drop this node's data until it's ready to send fresh data", like
upgrade (no timeout).