Optimize fs resource status on Linux
The agent verifies that the fs is not layered on a multipath with no
active path and queueing active.

This test was inefficient as a couple of dmsetup is triggered for each fs
underlying device.

This patch add a session cache for the dmsetup commands ls, table and status.

Also use a lazy for the dm major.