Implement a delta-replication log for node-to-node datasets exchanges
* Bump the monitor thread loop period to 30 seconds, and generalize the
wake up notification
* Add the json_delta lib to the project (BSD licenced)

Each change to a node dataset triggers a generation increment, and the
dataset diff is added to a log hash indexed by generation.

The log hash is purged from generations older than the min generation
amongst all reachable nodes.

If a node joins or come back online, it advertize either a 0 gen (init), or
a gen not in the log of the dataset responsible node. In this situation,
the full dataset is sent for resync, and subsequent exchanges are delta-based.