Refresh the services object cache on node.conf change
The services may cache references changed.

One such example is the join of a new node in a cluster with services
configured with nodes={clusternodes}. In this case, the cached service
object does not include the new join node until daemon restart or a
service configuration change event.

This patch rebuilds all service objects on node config changes.