Fix the reverse deprecated keywords handling
The nodedict sets a list of reverse deprecated keywords. For one
"asset_env" has ancestors set to "environment" and "environnement".

The extconfig module did not support lists, so this patch adds this
feature.

This fixes a stack in the monitor thread when the node env was not
set. In this case, deprecated keywords were tried and hit the
missing feature issue.