Fix relay hb oscillations
Don't trust the node data "updated" as a heartbeat last write indication, as
the written node data can be older due to the longer (30s) monitor loop delay
when no change mandates a node data refresh.

This patch adds a "updated" key to the data stored by the relay. The relay
is responsible for setting the value to the current timestamp on recv.

Then the relay serves the node data (encrypted) along with the "updated"
timestamp (unencrypted).