Fix disk hb oscillations
Don't trust the node data "updated" as a heartbeat last write indication, as
the written node data can be older due to the longer (30s) monitor loop delay
when no change mandates a node data refresh.

This patch adds a "updated" key to the data written on disk, with value set
as the current timestamp on write.

Then the peers rx threads read the node data (encrypted) along with the "updated"
timestamp (unencrypted).