Drop --name from a container.docker resource run_args
The agent already sets <svcname>.<rid> as the container's name to be able
to identify it.

We may want to switch to using label for that, but most docker versions
in the wild do not support label updates, so delay this change. In the
meantime this patch keeps users from shooting themselves in the foot.