Split the systemd opensvc-agent unit
Split a oneshot service out of the opensvc-agent, leaving it only in charge
of stopping and starting the daemon, never touching the services.

* Move the pushasset as a the new service ExecStart
* Move the daemon shutdown as a the new service ExecStop
* Set the oneshot service After to opensvc-agent.service.

This reorganization limits the risk of accidentaly stopping the services
when using systemd to stop the daemon. And remove black magic in the Node
class to guess if a daemon stop or daemon shutdown was intended.