Support zfs dataset in legacy mountpoint mode
Don't set mountpoint=<mnt> on start for datasets with mountpoint=legacy.

Instead switch to the system's mount/umount commands.

Keep setting mountpoint if its value is not legacy but is different from the
expected <mnt>.

This patch allow the agent to handle datasets in local zpools, replicated
with zfs sent/recv, without standby=true.

In legacy mode, the recv does not end up with the received datasets mounted.