Add the "status_log" app resource keyword
This boolean, if set, redirects the app checker stdout to the resource status
info-log, and stderr to warn-log.

Can be used to hint the cause of status degradation and make for faster
troubleshoots.

Beware not to report ever-changing outputs, as to not stress the cluster data
synchronization with futile changes.

Example:

$ ha1 print status -r
ha1                                down       warn
`- instances
   |- nuc                          undef      daemon down
   `- aubergine                    up         warn, idle, started
      |- ip#0              ....... up         128.0.0.2@lo
      |- ip#1              ....... up         128.0.0.3@lo
      |- ip#2              ......S stdby up   128.0.0.4@lo
      |- app:g1                               //
      |  |- app#1          ...../. up         trigger
      |  |                                    warn: bar
      |  |                                    warn: baz
      |  `- app#2          ...../. up         app
...