Fix keyword default value being altered because of reference-passing
Use a copy of the keyword default value as the OptNotFound.default instead
of a reference to the keyword default, so if the catcher then modifies the
value fetched the exception default, this does not change the keyword default.

This fixes resources tags corruptions, causing a start action on a service
with zoned filesystem to not match any resources.