Add a app.simple resource driver to handle non-daemonizing programs
Split the app resource driver into app.forking (the default) and app.simple.
A app.simple resource is as simple as:

[app#0]
type = simple
start = /usr/sbin/tcpdump -i lo

In this case:

1/ the checker defaults to a pgrep + validation of markers in the
process environment. A check keyword can be set to customize this default
behaviour.

2/ the stopper defaults to a kill -TERM on the process found by the default
checker. A stop keyword can bee set to customize this default.

For now the logs are not handled, but users can do their own logger pipes
or redirections in the start command.