Fix wrong return code when catching Svc excError from Node::do_svcs_action()
In this case the error was not translated into a non-zero return code.

For example a "svcmgr thaw --local" on a service the node is no longer
a candidate raises excError, but the return code was zero, and the daemon
who submitted this command would loop over this command again trying to
reach the thawed state in vain.