Add the user, group and cwd keywords to the app driver
If the binary is owned by root and user is set, demote to that user before
running the binary.

If the binary is owned by root and user is set and the group is set, demote to
that group before running the binary.

If cwd is set, use that as the working directory instead of the default
<pathtmp>.