Machine scope tuning enhancements
1/ Fix a potential conflict between vm names ending with the same
suffix. And simplify the scope file path generator.

2/ Add a function the remove the machine scope customization
file and hosting dir (if empty). Plug it in the container stop()
codepath

3/ Make sure the machine scope customization is incrementally added
on a start action when the container is already started. So if a
node is upgraded to this version of opensvc and has already
runnnig containers, a 'svcmgr start --rid container' will create
the customizations.