Fix drbd standby resource not rollbacking to the secondary role
Standby resource are never rollbacked but a drbd resource startstandby
promotes the role to primary, so the rollback should revert to secondary in
this case.

This patch adds a new resource flag "rollback_even_if_standby" defaulting to
False. The drbd driver sets it to True and implement a special rollback
method which cares about the role change on start.