Add "scale_up" and "scale_down" events
And make sure the nodes not in the service peers list are not seen as
candidates by the daemon. This can happen if the same service is deployed
on different cluster segments.

Also avoid clearing the status cache when detecting a service config change
but the status.json mtime is already > config mtime. This scenario is most
frequent because "set" and "edit config" actions refresh the status.json
immediately, before waking the monitor.