Don't refresh the status.json on cfg changing actions at the end of the action
Producing a correct status.json require rebuilding the service, so we can't
refresh the status.json from the same run.

The daemon detects the status.json is older than the config, and refresh the
status.json in this case.

Also don't set the local expect to "started" for scoped start. The daemon
will notice later if the local expect needs to be upgraded.