Daemon monitor thread fixes
1/ Change the "init" nmon state handling

Starting the daemon now immediately executes a "svcmgr -s '*' status --parallel
--refresh'.

The init period ends when this exec terminates.

During init, the get_services_status() is inhibited.

2/ Fix the startstandby resources

Some drivers (drbd) implements a specific startstandby method, so the
resources orchestrator should not execute a start but a startstandby action
when needed.

Also stop sending a start action to standby resources on started services
if the restart keyword is not set. If the user wants a restart this keyword
can be set.

3/ Harden the rejoined conditions

If the heartbeat still did not fill the "services" node data key, don't
consider the node joined. The node monitor key is filled before the service
data through the "ping" messages, but we need to wait for node service data
before starting to orchestrate.