Fix systemd reaping the block devices in parallel with opensvc shutdown
Add more Wants and After to the opensvc-agent service unit, notably the
blk-availability service which is calling blk-deallocate on dm and md block
devices.

One of the consequences of not sequencing opensvc shutdown before blk
deallocate is that the disk heartbeat is lost too soon and if it is the
only configured or beating beating heartbeat, the peers can takeover
before the opensvc services are shutdown.