Add the boot service action
This action will be triggered only by the node boot manager, before the
daemon starts up.

It is responsible for bringing down resources that either the system
activated but should not have (linux vg whose tag allowed the system to
activate), or resources that state persisted up after a crash-boot
(scsi reserv).