Implement the "svcmgr move --to <node>,<node>" async action
And rebase the switch and takeover action over this generalized way of moving
service instances.

The target status displayed in svcmon is "placed@", the "@" differentiating
from the "placed" target used by the "giveback" action. The full global_expect
contains the target nodes after the "@", but svcmon trims it for output
density optimization. The json data contains the full information though.