Add support for ipvlan in the ip.docker driver and the "mode" keyword
mode = ipvlan-v2 or ipvlan-v3 activates the ipvlan modes.

mode = dedicated is now the preferred way to activate the dedicated netif mode,
though tags = dedicated is still supported for backward compat.

This patch also adds a default gateway though the guest dev, as a fallback
when the gateway ip is not specified.

Of note:
* ipvlan requires a 4.2+ Linux kernel
* ipvlan modes are verified to work with OVH ip failovers