Smarter frozen peer instance merging on daemon startup
Previously, on daemon startup, any frozen remote instance would provoke the
freeze of the local instance. This to prevent starting the local instance
if the service was frozen while the local daemon was down.

This simple approach doesn't help with the "freeze instances on one node and
reboot the first node" scenario, as the administrator expects the instances
to be started on the rebooted node after the reboot.

This patch implements a smarter condition of the peer instance frozen state:

* Everytime the daemon is stopped (on unlock) update the <pathvar>/last_shutdown
file

* On startup, merge the peer instance frozen only if its monitor state changed
while the daemon was down