Change the "frozen" node and svc instance attr from bool to the freeze timestamp
This is must more informative than the boolean and not a big deal as
'if data.get("frozen")' still evaluates as True if a timestamp is set.

This patch also changes the merge_frozen() to use this timestamp instead of
the "status_updated" smon attribute.

Rename the merge_frozen() method into a more explicit merge_service_frozen().

Add a merge_node_frozen() method to freeze a starting node if a peer node has
been frozen while we were down.

And wrap merge_service_frozen() and merge_node_frozen() into a new
merge_frozen() method.