Detect if a scsi dev rescan is necessary on promote_dev_rw() on Linux
blockdev --getro <dev> can return dev is not readonly, but its scsi device
might still be readonly. A simple open(<dev>, O_NONBLOCK|O_RDWR) fails in
this case.

Use that as a test to determine if a scsi dev rescan is necessary after
we made sure the bdev is not readonly.