Add a pausable property to the sync drivers
Drivers handling hardware-assisted disks replication are set non-pausable,
other drivers default to pausable.

If pausable and the aggregated service status is not up, the resource status
is n/a.

If not pausable, always evaluate the real resource status.

This patch fixes the undue failback for a service with rollback=false an
instance of which failed to start in the resource just next after a
srdf resource. In this situation, the instance availstatus is still down
because the srdf resource status is n/a due to the paused mecanism, instead
of up.