Enhancement for "nodemgr updatecomp"
Up to this patch, compliance designers were required to ship
a clone of <var>/compliance/com.opensvc/ in there compliance
modules archive, because the whole content of this directory
was overwritten by the archive content, so the upstream
com.opensvc/ was lost.

This behaviour also messes with the opensvc package checksums,
for package managers that care about that (rpm for one).

This patch makes sure that the existing com.opensvc/ is still
available unchanged after a "nodemgr updatecomp" IF the compliance
archive does not contain a com.opensvc/ already. In this later
case, we still install the com.opensvc/ from the archive, as was
already did before.