Make json config feeding through stdin explicit, via --config=<-|/dev/stdin>
"svcmgr create" accepted json config feeding through stdin if stdin was not
a tty.

This caused a hang with "cat <conf.json> | ssh n2 svcmgr -s <svcname> create"
because ssh does not allocate a pty by default on the remote node, and we
ended up reading from stdin and never getting any data.

This patch replaces the "not isatty" test with a test on the --config value,
which now must set to "-" or "/dev/stdin" to trigger stdin reading.

And adapt the svcmgr create command, forged by the daemon for scaling, to set
"--config=-".