Report n/a instead of up for scsireserv status if a dev is ro
Because in this case, we can't clear the reservation until the dev is
promoted rw.

This happens on srdf devices that became r2 after a failover due to a crash.
When the crashed node comes up again, the reservation are still held on the
r2 devices, and they can't be dropped.

Still report the situation as a resource log "info" message.

Example:

     |- disk#0pr         ...O... n/a       /dev/sdat, /dev/sdau, /dev/sdcp, /dev/sdcq,
     |                                      /dev/sdem, /dev/sden, /dev/sdgj, /dev/sdgk
     |                                      info: resv held on ro dev /dev/sdgj