Force the same session uuid on detached actions
Delete, unprov, stop,... actions run in a forked process so they can finish
properly even when the parent process gets killed by a "fuser -k".

So we must take care of forcing the same session uuid in parent and child to
be sure cache updates/clears from the children are seen from the parent, which
may submit follow-up actions.

This bug caused some nosetests doing prov (cache1)/unprov(cache2)/prov(cache1)
to fail, the 2nd prov working with a unsynced cache (cache1) because the unprov
cleared cache2 instead of cache1.

This patch verifies a OSVC_PARENT_SESSION_UUID is present in the environment,
and if so uses that instead of generating a new uuid.

OSVC_PARENT_SESSION_UUID is set by the svc detached action wrapper function in
svcmgr.