Add a scopable "disable" keyword in all hb sections of node.conf
The default value is False.

This settings is used to not-start and stop the hb threads, which is better
than letting svcmon display alarming information like "thread stopped", or
"peer unreachable", when the situation is as-designed.

For example, a disk heartbeat configuration can only apply to a cluster
segment (n1,n2) where the disk is visible. The out-of-segment nodes should not
report alarming information for this heartbeat.

Configuration for this example:

[hb#4]
type = disk
dev = /dev/mapper/36....
disable@n1 = false
disable@n2 = false
disable = true