Add the "toc" action to the list of action mandating slave relaying
When monitor action is set to "freezestop", the stop() codepath is executed
with the action name set to "toc".

When the service has encap resources and a container in noaction (for sharing
with other services), the encap part needs to be stopped for the instance to
return to a "down" availstatus and thus not block the failover.