Add a "nodes" keyword to all heartbeats to allow the hb on a cluster segment
For example, a disk heartbeat visible only one 2 out of 4 nodes:

[hb#4]
type = disk
nodes = n1 n2
dev = /dev/mapper/36...

With this new keyword, the hb driver can now limit the slot lookup and reading
to the specified nodes. And the peers reported in the hb status data are also
limited to the specified nodes.