Add support for services namespaces
Namespaces allow users to have services with the same name in different
namespaces.

The svcmgr and svcmon service selector (--service option) now accept
service paths (svcpath) in the form: <namespace>/<svcname>.

Users have 3 methods to select a namespace to operate into:

1/ svcpath selector expressions (ex: svcmgr -s test/svc1 create)
2/ --namespace=<namespace> (ex: svcmgr -s svc1 --namespace=test create)
3/ OSVC_NAMESPACE environment variable (ex: OSVC_NAMESPACE=test svcmgr -s svc1
create)

Exporting OSVC_NAMESPACE=test thus applies all svcmgr and svcmon action in
the context of the test namespace.

Operating services through their svcmgr symlink is not supported for services
in namespaces, to avoid relying on the PATH env var paths ordering. Operating
via svcmgr is thus mandatory for these services.

A new {namespace} reference is now available.

Service created before this patch and services created without an explicit
namespace are considered in a pseudo namespace named "root". Their config are
stored as usual in /etc/opensvc/.

Service in an explicit namespace have their config stored in
/etc/opensvc/namespaces/<namespace>/. They also have the namespace name
inserted in their cgroup path and in docker container names on Linux.

Children and parents relationship can now use svcpath. If a svcname is
declared in these lists, the same namespace is implied. To declare a
relation to a service in another namespace, users have to use svcpath.
Relations to services in the root pseudo namespace have to be defined as
root/<svcname> to avoid ambiguity.