Support svcmon node columns filtering/hiding
svcmon --node option now accepts a node selector expression.

If not specified, all nodes are selected, and the comm endpoint is the unix
socket.

If specified, only matching node columns are displayed, and the comm endpoint
is the unix socket if the local node is matching, or the first matching node
as a fallback.