Add a cluster lock infrastructure
Locks are acquired and released via the unix socket listener.
Acquire needs a lock name and timeout (default 10s), and returns a lock_id.
Release needs the lock name and lock id returned by acquire.

Acquire first try to claim the lock, which can fail if an older claim or
accepted lock with the same name is there.

Once claimed locally, the lock claim is propagated through the heartbeats,
and the listener client thread waits for all live nodes to accept the claim,
at which point the lock id is returned to the caller.

A node receiving a peer dataset will janitor its local view of the cluster
locks:

* don't re-merge a released lock emitted by the janitoring node
* merge if no local lock by that name
* merge if there is a local lock by that name, but the peer lock is older
* flush local locks that were released on their originating node

Use this cluster lock infrastructure to garanty the freenas pool won't
create a disks concurrently (for the free aligned lun algo).