Stricter node and cluster mirroring on nodemgr daemon join
Strictly mirror the joined node/cluster dispatch of keywords on the joiner.

And don't migrate cluster.nodes to system/cluster on the joined node if the
user choosed to set it up in node.conf.

This patch changes the "join" lsnr handler returned structed format, to
separate clearly the node/cluster data.