Add more service selector capabilities
New syntaxes:

- **
  select all objects

- test/**
  select all objects in the test namespace whatever the kind

- **/test*
  select all objects named test* whatever the namespace and kind

Also strip svcpaths in svcmon and svcmgr user-friendly outputs to their
shortest non-ambiguous form, using "svc" as the default kind:

- "prod/svc/svc1" is the shortest form if no namespace is selected, and
"svc1" is the shortest form if the "prod" namespace is selected.

- "prod/vol/svc1" is the shortest form if no namespace is selected, and
"vol/svc1" is the shortest form if the "prod" namespace is selected.

Use "svc" as the default kind filter for svcmon and svcmgr. Use "**" to
list all objects.