Disable libvirt resource partitions if machinectl is installed
Some distributions ship machined pumbing in the core systemd
package, and make it very hard to disable.

Libvirt doesn't fallback to manual cgroup management if
systemd machined is installed, event if it does work properly
when we feed it our partition, non-parented to machine.slice.

Better to disable the partition logic in this case, to not
fail starting the container, even if this leads to the container
not being parented to the opensvc service cgroup.