Honor timeout in the sending part of daemon_send()
And set timeout to most node daemon_send() callers.

On Solaris we consider errno 146 as a retyable, as this error has been seen
emited in situations where the daemon is running. So timing out is the last
option to not block node actions like ping, wake_monitor, etc when the
daemon is shutdown and not going up again.