Add a 10 seconds timeout to the threads join on daemon shutdown
Each thread janitors a list of sub threads that it tries to join() when
asked for stop. Don't block the daemon shutdown if some of these sub thread
refuse to go away in a reasonable delay.

This patch also adds the alive listener thread info to the listener stats,
and dumps the head thread status() as flat_json in the logs when the
sub-threads join times out, so we have a trace of which listener
sub-thread is hanging around.

$ om daemon status --format json | jq .listener.stats.sessions.alive
{
  "1d1ac5a1-f2ac-4f56-9c6f-3d7697df7cc2": {
    "addr": "local",
    "encrypted": false,
    "progress": "sending action_daemon_status result"
  }
}