Add cfg objects
Configurations are objects very similar to secrets, but not limited to
shm-backed volumes, and with data keys not encrypted.

Literal values are prefixed with "literal:", file contents are base64-encoded
and profixed with "base64:".

The "secrets" keyword is moved from docker/podman containers/tasks to
volumes. Now containers can reference secrets via "volume_mounts".

The "configs" keyword is added to volumes. The "config_environment" keyword is
added to docker/podman containers/tasks.