Allow async actions on just created services
The global_expect validation raised with "service not found" for async
actions submitted just after create (before the instances get loaded in
the cluster data).

This patch make the validator wait for 5s max for the service to appear
before giving up with a service not found (if the service config file
exists, else it fails immediately).