Remove the sync.zfs can_sync() check on snap timestamp youger than schedule interval
This test is wrong as the osvcd scheduler already did this check but based on
the last run timestamp, which is older than the snap timestamp, so the
can_sync() would return False and the sync won't be retried by the scheduler
until the next schedule allowed period.

The visible effect was commonly a daily sync being executed every odd day.