Fix "om node wait" stack and support waiting on key changes outside .monitor.nodes
The previous patch opened wait for key changes outside .monitor.nodes, but the
current code applied node data diffs only, so the initial cluster_data was not
updated for keys not in .monitor_data.nodes.

This patch reverts to fetching a new cluster_data upon patch events, which is
slower but necessary.