Fix the "encap" value in instance status data
When a encap service has a container resource (docker for example), the
encap status data "encap" key was {} instead of True.

This caused the collector to add a extra line for the encap instance in the
svcmon table.

Also catch the TypeError in the instance status renderer when accessing
the encap subkeys, so the encap=True case is correctly catched.