Create <pathtmp> only on first use
Instead of creating it on rcGlobalEnv import.

Because rcGlobalEnv is imported from almost every code path, including
the "om svc print config mtime" one, which is called between zone boot and
zone reaching the multi-user stage. If the /var/tmp/opensvc dir is created
in this period, the dataset mount on /var/tmp will fail because the mount
point is no longer empty.

This is also benefiting performance, saving a stat() and utime() when not
needed.