Relay and multiplex a listener action to peer nodes
The tls listener usually listens on a vip floating on cluster nodes, so
CRM-level actions submitted to the listener may not land on the targeted node.

This patch allows requesters to set a node selector in the request, like:

{
    "action": "service_action",
    "node": "region=fr",
    "options": {...}
}

This CRM level action will be relayed by the tls listener router to all nodes
matching the selection, and aggregate the results in a dict indexed by node.