Client side support for CRM level commands
On node and on svc/vol/sec/cfg/usr objects.

This patch refactors the "create" action codepath for all type of inputs
(template, config file, selector, stdin) produce a common dataset: a dict
of definitions indexed by object path.

This dataset can thus be fed to the tls socket or installed as a local file,
depending on the "context".

This patch also adds support for --config=<fpath or uri to a json content>,
which extends the --config=<fpath or uri to a ini content>.