Add a special create action multiplexer
A create action submitted to the tls listener should be routed to one of the node
of the object's node selection (DEFAULT.nodes).

With this patch, "om <path> edit" can update the configuration of an object that
has no instance on the remote cluster node that handles the tls connection.