Fix latent issue with duplicate kwargs in the lsnr handlers
The request router sets its own kwargs then merges the request "options" dict
in the same kwargs, causing possible clash.

Pass options to handlers as a dict.

This patch makes use of this fix to permet service_action to receive "action"
in the options, and allow "get" and "eval" to namespace guests.