Automatically clear "start foiled" for services avail up
If a daemon disappears, a secondary node may decide to start a service instance
for a service that has a up instance on the node with the dead daemon.

This start can fail on resource conflict (example public ip), and the instance
is flag "start failed" to avoid retry.

When the dead daemon comes up again, the service will be up avail up, but
secondary instance keep the "start failed" flag, preventing further failover.

This enhancement is notably import for the cluster vip service.