Change the flex orchestration logic
Add a flex_target keyword, used by the orchestrator as the number of
instance that should be started.

Its integer value defaults to the flex_min value, so a cluster upgrading to
this version would start the same number of instances, but excess instances
would be stopped (previous version would only stop if >flex_max, now stops
if >flex_target).

This patch also deprecates flex_min_nodes and flex_max_nodes in favor of
simpler flex_min and flex_max. Compatibility is maintained.

This patch closes the following dead end:

* On a 2-nodes cluster
* A svc1 service has parents=flexsvc
* flexsvc has flex_min=2 because we want 2 instances started in optimal situation
* node1 is lost
* node2 is rebooted
=> svc1 would "wait parents" until node1 comes up again

This patch allows the user to setup:

* flex_target=2 on flexsvc, so 2 instances get started in optimal situation
* flex_min=1 (default), so the avail status is up if at least one instance is started
=> svc1 would start on node2 reboot even if node1 is lost