Add jwt authentication to the listener
If listener.well_known_uri is set, verify the Bearer token is
valid and signed by the openid provider.

The provider is in charge of setting the "grant" as a claim in
the token (string or list of string).

This patch also adds a /authinfo api handler, to let the webapp
known which openid provider it must redirect to to obtain tokens.