Fix "do <action> ... (<origin>)" not being logged anymore
These log lines are useful in troubleshoot, to determine the
origin of actions and the exact command being passed.

The collector also use that to distinguish master actions.

This feature was broken by the patch changing the argv[0] from
svcmgr.py to mgr.py, as we test this item to decide if we need
to log the line.

Closes #7