Add the "reset_options=true|false" boolean keyword to sync.rsync
The sync.rsync driver defines a set of default rsync options,
including -A and -X if the src system supports them.

These options could not be switched off, which is problemtic at
least when the default options trigger a bug like unallowed
selinux security labels xattr reset.

This patch adds a "reset_options=true|false" keyword, defaulting
to false. If set to true, the options defined in "options" are
used directly instead of concatenated with internal default
options.