Fix the ip drivers dns_wait behaviour
The dns_wait=10 ip definition was handle by waiting for a
dummy condition to become true using the GET /wait handler on
each service peer.

The bogus condition aside, this method is not efficient.
Replace it by a GET /wait call waiting for ...<rid>.info.addr to
be set, the a GET /sync call to wait for the propagation of
the dataset.