Protect the toc action with the object action lock
If monitor_action is not "reboot" or "crash".
So "toc" can not run twice in parallel.

Before this patch, a toc action with monitor_action=freezestop could run in
parallel on the same object, causing a race to stop the resources, and unclean
stop (fuser -k to free fs while app resources are still stopping, for ex).