Add the "enter --rid <rid>" action
Implement for docker and podman first.
Detect a shell in the container and execute interactively with a tty.
This action avoids having to use "oci exec -it {n} bash" syntax.

Also add a GET /object_enter api handler, launching a singe-use gotty server,
with random creds, random port and random path. A browser has then a few
seconds to connect before it dies.