Handle path returning "Not ready" scsi errors on scsi3 pr inquiries
Huawei Dorado and EMC SRDF Metro set the paused member lun of the
replication set in such a mode.

In this case, we still want to be able to prstop and successfully stop a
service instance, so these errors must be ignored.

Still log warnings about this situation.

Also warn in resource status when the number of registered paths does not
match the number of paths.