Add mpathpersist support
Detect mpathpersist presence and prefer it over sg_persist.

The sg_persist is then no longer necessary, the codebase using mpathpersist
for all commands.

OpenSVC doesnt set the defaut.reservation_key=file in multipath.conf. This
is important to set it to activate the register-path-before-add-to-multipath
feature of multipathd, which prevents ioerrs on registered multipath when
one of the active paths is not registered.

Even if defaut.reservation_key=file is not set, the mpathpersist commands we
execute produce the same registration and reservation as with sg_perist.