Allow empty sec and cfg keys value
It might be valid to expose an empty environment variable to a container.

Also it is frequent to find empty files in a file tree imported in a sec or
cfg. In this case, before this patch, the "om <secpath> add --from <head>"
failed when trying to add the key with empty value, leaving the file tree
partially imported.