Move adder functions from svcBuilder to each driver module
The adder functions translate the resource configuration in ini format to
the driver class kwargs.

Moving them to the driver modules allow adding drivers without changing the
svcBuilder core file. This is the first requirement of out-of-tree drivers.

This patch also applies a strict policy on driver names, so mimport can
always be used to lookup the driver modules from rgrp/rtype.

Also some sync classes did not begin with a uppercase, in violation to pep8.