Move commands to the 'commands' package
And command option parsers to command.<command>.parser

Add a __main__ and __init__ to the lib/ directory, so lib is now a proper
exectable package.

Adapt bin/om and bin/svcmgr to use "python -m lib $@" to execute opensvc
commands.

Add a --foreground option to "om daemon start", so users that want to run
the daemon in foreground don't have to know the python magic formula.