Remove adder functions from drivers
Use the kwstore to format resources kwargs instead, from the builder.

This moves all the pre-resource-instanciation magic to the drivers themselves,
which is also a postulate mandated by out-of-tree drivers.

Rewrite bin/pkg/make_doc in python, and remove the __main__ from dictionnaries.

Add a load_drivers() function to utilities.drivers

Add a protoname atribute to Keyword, defaulting to the keyword name. Use this
attribute when the kwarg name is different from the keyword name in the config
file.