Add a verify_dev() implementation to linux disk.raw
Make it verify a loop is actually linked to a file.

This fixes a case of the disk.raw resource referencing a loopfile still
reported as up after a start, because even if the loop was unconfigured
the exposed_devs[0] is still known as loop12 (for ex) and /dev/loop12
still exists.