Add a core.capabilities module
Capabilities are a list of tags describing the node characteristics.

A capability must be long-lived, as we refresh capabilities only on
daemon restart and on system package uninstall/install.

Two new commands are introduced:
* om node print capabilities
* om node scan capabilities

The capabilities hold a cache of available commands for example, which
allow drivers not to resort to which(), ie slow syscalls, to determine if
a command is available.

This patch converts a first batch of which() uses to capabilities.

Drivers can contribute capabilities too. It will become mandatory for a
driver to implement the driver_capabilities() function that will announce
if the driver requirements are met, via the
drivers.resource.<DRIVER_GROUP>.<DRIVER_BASENAME> capability.

A driver can also insert additional sub-capabilities under its namespace.
For example, this patch converts the disk.scsireserv driver to load its
preference between mpathpersist and sg_persist via such sub-capabilities.